requirements.txt
pandas>=2.0.0
mysql-connector-python>=8.0.0
python-dateutil>=2.8.2 